src-tauri/resources/memory-service/pyproject.toml
[project]
name = "orbit-memory-service"
version = "0.1.0"
description = "Memory service for Orbit using mem0 + FAISS"
requires-python = ">=3.10"
dependencies = [
    "mem0ai>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "faiss-cpu>=1.9.0",
    "pydantic>=2.0.0",
]

[project.scripts]
serve = "server:main"
